Format notification for telegram with its title and add info type
